Use first 16-bit timer for wallclock timing

+++ timer.h
@@ -3,5 +3,5 @@
 #include <stdint.h>
 
 void timer_init();
-void timer_wait_for_ms(uint64_t wait_millisec);
-void timer_wait_until(uint64_t wakeup_millisec);
+void timer_wait_for_ms(uint32_t wait_millisec);
+void timer_wait_until(uint32_t wakeup_millisec);
